Removed Communicator, TestClient and TestServer
Tidying up
