+++ src/main/java/commons/rpc/ClientCommunicator.java
@@ -12,7 +12,16 @@ import java.nio.ByteBuffer;
 
 // To use, run TestServer on 1 terminal then run TestClient on another terminal
 
-public class ClientCommunicator extends Communicator {
+public class ClientCommunicator {
+    int clientPort;
+    InetAddress serverAddress;
+    int serverPort;
+    int packetSize;
+    int messageSize;
+    int headerSize;
+    byte[] buffer_req;
+    DatagramSocket socket;
+    int requestID;
 
     /**
      * Creates a socket to send / receive UDP packets at specified port number
@@ -20,25 +29,68 @@ public class ClientCommunicator extends Communicator {
      *
      * @param port port number for UDP socket
      */
-    public ClientCommunicator(int port) {
-        super(port);
+    public ClientCommunicator(int clientPort, InetAddress serverAddress, int serverPort) {
+        this.clientPort = clientPort;
+        this.serverAddress = serverAddress;
+        this.serverPort = serverPort;
+        this.requestID = 0;
+        this.packetSize = 512;
+        this.headerSize = 16;
+        this.messageSize = this.packetSize - this.headerSize;
+        try {
+            this.socket = new DatagramSocket(clientPort);
+        } catch (SocketException e) {}
+        /**
+         * 4 bytes for requestID
+         * 4 bytes for the datagram number
+         * 4 bytes to send the total number of datagrams.
+         * 4 bytes to send length of message in datagram
+         */
     }
 
-    /**
-     * Sends request / response to destIP and destPort
-     *
-     * @param o        request / response object
-     * @param destIP   destination IP
-     * @param destPort destination port
-     */
-    @Override
-    public void send(Object o, InetAddress destIP, int destPort) {
-        if (!(o instanceof commons.requests.Request)){
-            throw new Error("Client communicator only sends Request");
+
+    public void send(Request r) {
+        ByteBuffer dataBuf = ByteBuffer.allocate(2000);
+        Serializer.serializeObject(r, dataBuf);
+
+        int totalDatagramPackets = (int) Math.ceil(dataBuf.position() / (float) this.messageSize);
+        int dataBufPtr = 0;
+        int dataBufMaxPos = dataBuf.position();
+
+        for (int i = 0; i < totalDatagramPackets; i++){
+            ByteBuffer packetBuf = ByteBuffer.allocate(this.packetSize);
+            packetBuf.putInt(this.requestID);
+            packetBuf.putInt(i);
+            packetBuf.putInt(totalDatagramPackets);
+            byte[] packetByteArray;
+            dataBuf.position(dataBufPtr);
+
+            if (i == totalDatagramPackets - 1){ // last packet or when only 1 packet needed - only writes remainder of dataBuf
+                int lengthRemaining = dataBufMaxPos - dataBufPtr;
+                packetBuf.putInt(lengthRemaining);
+                packetByteArray = packetBuf.array();
+                dataBuf.get(packetByteArray, this.headerSize, lengthRemaining);
+            }
+            else{ // writes messageSize worth of data to DatagramPacket
+                packetBuf.putInt(this.messageSize);
+                packetByteArray = packetBuf.array();
+                dataBuf.get(packetByteArray, this.headerSize, this.messageSize);
+                dataBufPtr += this.messageSize;
+            }
+            DatagramPacket message = new DatagramPacket(packetByteArray, packetByteArray.length, this.serverAddress, this.serverPort);
+
+            try {
+                this.socket.send(message);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
-        else{
-            super.send(o, destIP, destPort);
-        }
+        /**
+         * comment out row below to test duplicate request
+         */
+        this.requestID += 1;
+
+//        System.out.println("Sending message" + " to " + destIP + " " + destPort);
     }
 
     /**
@@ -69,5 +121,31 @@ public class ClientCommunicator extends Communicator {
         System.out.println(((TestResponse) response).testString);
 
         return response;
+    }
+
+    protected Packet receivePacket(){
+        byte[] buffer = new byte[this.packetSize];
+        DatagramPacket message = new DatagramPacket(buffer, buffer.length);
+
+        try {
+            this.socket.receive(message);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        ByteBuffer received_bb = ByteBuffer.wrap(buffer);
+        int requestID = received_bb.getInt();
+        int datagramNum = received_bb.getInt();
+        int totalDatagramPackets = received_bb.getInt();
+        int messageSize = received_bb.getInt();
+
+        InetAddress senderAddress = message.getAddress();
+        int senderPort = message.getPort();
+
+        ByteBuffer messageBuffer = ByteBuffer.allocate(messageSize);
+        messageBuffer.put(buffer, this.headerSize, messageSize);
+        messageBuffer.position(0);
+
+        return new Packet(requestID, datagramNum, totalDatagramPackets, messageSize, senderAddress, senderPort, messageBuffer);
     }
 }
